chore: header styling iteration (#522)
