+++ package.json
@@ -110,7 +110,6 @@
     "react-error-boundary": "^4.0.12",
     "react-router-dom": "^6.22.2",
     "react-toastify": "^10.0.5",
-    "react-tooltip": "^5.26.3",
     "styled-components": "^6.1.8",
     "svg-inline-loader": "^0.8.2"
   },
