+++ src/renderer/library/Menu/Wrapper.tsx
@@ -5,8 +5,8 @@ import styled from 'styled-components';
 
 export const MenuWrapper = styled.div`
   position: absolute;
-  right: 0.75rem;
-  top: 2.75rem;
+  right: 0;
+  top: 2.25rem;
   z-index: 10;
 
   display: flex;
@@ -14,7 +14,7 @@ export const MenuWrapper = styled.div`
   justify-content: flex-start;
   width: 175px;
   height: auto;
-  padding: 0.45rem 0;
+  padding: 0.75rem 0 0;
 
   box-shadow: -1px 3px 5px 0px var(--card-shadow-color);
   border: 1px solid var(--border-mid-color);
@@ -43,6 +43,60 @@ export const MenuWrapper = styled.div`
     color: #636363;
     &:hover {
       background: inherit;
+    }
+  }
+
+  // Controls
+  .controls {
+    .controls-wrapper {
+      margin-top: 0.75rem;
+      border-top: 1px solid #282828;
+      border-bottom-left-radius: 0.75rem;
+      border-bottom-right-radius: 0.75rem;
+      background-color: #1b1b1b;
+
+      padding: 1rem;
+      display: flex;
+      flex-direction: column;
+      row-gap: 1rem;
+      width: 100%;
+
+      .menu-btn {
+        background-color: var(--background-menu);
+        border: 1px solid var(--border-mid-color);
+        font-size: 1rem;
+        font-family: InterSemiBold, sans-serif;
+        user-select: none;
+        transition: background-color 0.2s ease-out;
+        width: 100%;
+        color: var(--text-color-secondary);
+
+        &:hover {
+          background-color: inherit;
+          border: 1px solid var(--border-secondary-color) !important;
+        }
+      }
+
+      // Connect button
+      .connect-wrapper {
+        position: relative;
+        width: 100%;
+
+        .do-pulse {
+          animation: pulse 3s infinite ease-in-out;
+        }
+        @keyframes pulse {
+          0% {
+            opacity: 1;
+          }
+          50% {
+            opacity: 0.5;
+          }
+          100% {
+            opacity: 1;
+          }
+        }
+      }
     }
   }
 `;
